requirements.txt
python-jose[cryptography]
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
pymysql